+++ data/migrations/202002251241_mail_5_for_IT.sql
@@ -1,0 +1,83 @@
+UPDATE mails SET content='<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Transitional//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd">
+<html xmlns="http://www.w3.org/1999/xhtml">
+<head>
+<meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
+<title>SHARE''NGO</title>
+<link href=''https://fonts.googleapis.com/css?family=Lato:400,900'' rel=''stylesheet'' type=''text/css''>
+</head>
+
+<body style="text-decoration:none !important;">
+<table width="500" border="0" cellspacing="0" cellpadding="0">
+<tr>
+  <td width="500" height="45"><a href="http://www.sharengo.it" target="_blank"><img src="http://www.sharengo.it/images/mails/Sharengo_Header_LOGO_500x45.gif" width="500" height="45" alt="logo_sharengo" /></a></td></tr>
+  <tr>
+  <td  width="500" height="158"><a href="http://www.sharengo.it" target="_blank"><img src="http://www.sharengo.it/images/mails/VISUAL_profilo_sospeso.gif" width="500" height="158" alt="visual_sharengo" /></a></td></tr>
+  <tr><td width="500" height="20"></td></tr>
+  <tr>
+  <td>
+  <!-- TABELLA TESTO -->
+  <table width="500" border="0" cellspacing="0" cellpadding="0"><tr>
+  <td width="10"></td>
+  <td width="480" bgcolor="#FFFFFF" style="font-family:''Lato'', arial, sans-serif; font-size:14px; color:#333;  font-weight:regular; line-height:20px; text-align: justify;">
+  
+<p>Ciao %1$s,</p>
+<p>abbiamo dovuto sospendere il tuo account perch&eacute; ci risulta che uno o pi&ugrave; pagamenti non sono andati a buon fine.</p>
+<p>Per tornare ad usare le nostre auto controlla la scadenza della tua carta di credito o prepagata:</p>
+<ul style="color: #08bd53; list-style-type: disc;">
+<li><span style="color: #333;">se&nbsp;<strong>non &egrave; scaduta</strong>, ricarica la tua carta prepagata o assicurati di non aver superato il plafond</span></li>
+<li><span style="color: #333;">se <strong>scaduta</strong>:&nbsp;<br /></span><span style="color: #000000;">- accedi in Area Riservata e disattiva la carta scaduta cliccando su &ldquo;<strong>DATI DI PAGAMENTO</strong>&rdquo; e poi su &ldquo;<strong>CAMBIA CONTRATTO</strong>&rdquo; <br />- per inserire una nuova carta ed effettuare il pagamento ti baster&agrave; cliccare su &ldquo;<strong>DATI DI PAGAMENTO</strong>&rdquo; e poi su &ldquo;<strong>EFFETTUA PAGAMENTO</strong>&rdquo; sotto la corsa da pagare</span></li>
+</ul>
+<p>Se non riesci a risolvere da solo il tuo problema, scrivi a <a style="color: #08bd53; text-decoration: none;" href="mailto:servizioclienti@sharengo.eu" target="_blank">servizioclienti@sharengo.eu</a>. </p>
+<p>Insieme sistemeremo il tuo profilo per permetterti di tornare subito a bordo di una Share&rsquo;ngo!</p>
+<p>Grazie mille per la collaborazione. <br /><br /> Un caro saluto,</p>
+</td>
+<td width="10">&nbsp;</td>
+</tr>
+</tbody>
+</table>
+<!-- FINE TABELLA TESTO --></td>
+  </td>
+  </tr>
+  <tr><td width="500" height="20"></td></tr>
+  <tr>
+  <td  width="500" height="40" bgcolor="#FFFFFF" style="font-family:''Lato'', arial, sans-serif; font-size:12px; color:#333;  font-weight:regular; text-align:center;"><img src="http://www.sharengo.it/images/mails/Sharengo_Footer_TEAM_500x40.gif" width="500" height="40" alt="sharengo_team" /></td></tr>
+  <td  width="500" height="40" bgcolor="#FFFFFF" style="font-family:''Lato'', arial, sans-serif; font-size:12px; color:#333;  font-weight:regular; text-align:center;">
+  Info e regolamento su <a href="http://www.sharengo.it" target="_blank" style="color:#08bd53; text-decoration:none;">www.sharengo.it</a></td></tr>
+  <tr>
+  <td  width="500" height="80" bgcolor="#1c2429" style="font-family:''Lato'', arial, sans-serif; font-size:12px; color:#333;  font-weight:regular; text-align:left;">
+  <!-- TABELLA FOOTER -->
+  <table width="500" height="80" border="0" cellspacing="0" cellpadding="0">
+  <tr>
+  <td width="10"></td>
+  <td style="font-family:''Lato'', arial, sans-serif; font-size:12px; color:#08bd53;  font-weight:regular; text-align:left; vertical-align:bottom;">SEGUICI SU</td>
+  <td></td>
+  <td width="10"></td>
+  </tr>
+  <tr>
+  <td width="10"></td>
+  <td>
+  <!-- TABELLA SOCIAL -->
+  <table width="135" height="30" border="0" cellspacing="0" cellpadding="0">
+  <tr>
+  <td><a href="https://www.facebook.com/Sharengozd/" target="_blank" style="color:#08bd53; text-decoration:none;"><img src="http://www.sharengo.it/images/mails/Sharengo_Social_FB.gif" width="30" height="30" alt="sharengo facebook" /></a></td>
+  <td width="5"></td>
+  <td><a href="https://twitter.com/sharengozd" target="_blank" style="color:#08bd53; text-decoration:none;"><img src="http://www.sharengo.it/images/mails/Sharengo_Social_TW.gif" width="30" height="30" alt="sharengo twitter" /></a></td>
+  <td width="5"></td>
+  <td><a href="https://www.linkedin.com/company/share''ngo/" target="_blank" style="color:#08bd53; text-decoration:none;"><img src="http://www.sharengo.it/images/mails/Sharengo_Social_IN.gif" width="30" height="30" alt="sharengo in" /></a></td>
+  <td width="5"></td>
+  <td><a href="https://www.instagram.com/sharengoit/" target="_blank" style="color:#08bd53; text-decoration:none;"><img src="http://www.sharengo.it/images/mails/Sharengo_Social_INST.gif" width="30" height="30" alt="sharengo in" /></a></td>
+ 
+  </tr></table>
+  <!-- FINE TABELLA SOCIAL -->
+  </td>
+  <td style="font-family:''Lato'', arial, sans-serif; font-size:12px; color:#999;  font-weight:regular; text-align:right;">Servizio clienti <span style="color:#ccc;">02 87317140</span><br/>
+© Share’ngo CS Group Spa 2016</td>
+<td width="10"></td>
+  </tr>
+  </table>
+  <!-- FINE TABELLA FOOTER -->
+  </td></tr>
+  </table>
+ 
+</body>
+</html>' WHERE category=5 AND subject='Share’ngo: Profilo sospeso ';